+++ src/web/assets/frontend/src/js/captchas/recaptcha-enterprise.js
@@ -122,8 +122,8 @@ export class FormieRecaptchaEnterprise {
     }
 
     onValidate(e) {
-        // Don't validate if we're going back in the form
-        if (this.$form.goBack || this.$placeholder === null) {
+        // Don't validate if we're not submitting (going back, saving)
+        if (this.form.submitAction !== 'submit' || this.$placeholder === null) {
             return;
         }